Beter code

+++ Driver/Sources/Driver_ADC.c
@@ -2,7 +2,7 @@
 #include "Driver_ADC.h"
 
 
-void MyADC_Init(ADC_TypeDef *ADC, uint32_t channel) {
+void MyADC_Init(ADC_TypeDef *ADC, uint8_t channel) {
     if (ADC == ADC1) RCC->APB2ENR |= RCC_APB2ENR_ADC1EN;
     else if (ADC == ADC2) RCC->APB2ENR |= RCC_APB2ENR_ADC2EN;
     else return;
